Demonstration of a simple unsafe cookie auth
Demonstration of a simple unsafe way to do session-token based
authentication. Users register with a username and password, then can
log in with a username and password. The only page they can access is
the root '/' page, and if they don't have a valid cookie of name 'auth',
then they will be fed a 403 error.